Move string parsing logic to integrations.h for better modularity and testability
