+++ tracer/src/Datadog.Trace.ClrProfiler.Native/cor_profiler.cpp
@@ -1129,62 +1129,8 @@ void CorProfiler::InitializeTraceMethods(WCHAR* id, WCHAR* dd_trace_methods)
     // first make sure this works
     if (rejit_handler != nullptr)
     {
-        std::vector<IntegrationDefinition> integrationDefinitions;
-
-        shared::WSTRING dd_trace_methods_string = shared::WSTRING(dd_trace_methods);
-        auto dd_trace_methods_type = shared::Split(dd_trace_methods_string, ';');
-
-        for (const shared::WSTRING& trace_method_type : dd_trace_methods_type)
-        {
-            shared::WSTRING type_name = WStr("");
-            shared::WSTRING method_definitions = WStr("");
-
-            // [] are required. Is either "*" or comma-separated values
-            // We don't know the assembly name, only the type name
-            auto firstOpenBracket = trace_method_type.find_first_of('[');
-            if (firstOpenBracket != std::string::npos)
-            {
-                auto firstCloseBracket = trace_method_type.find_first_of(']', firstOpenBracket + 1);
-                auto secondOpenBracket = trace_method_type.find_first_of('[', firstOpenBracket + 1);
-                if (firstCloseBracket != std::string::npos &&
-                    (secondOpenBracket == std::string::npos || firstCloseBracket < secondOpenBracket))
-                {
-                    auto length = firstCloseBracket - firstOpenBracket - 1;
-                    method_definitions = trace_method_type.substr(firstOpenBracket + 1, length);
-                }
-            }
-
-            if (method_definitions.empty())
-            {
-                continue;
-            }
-
-            type_name = trace_method_type.substr(0, firstOpenBracket);
-
-            Logger::Info("InitializeTraceMethods:  Considering method definitions ", method_definitions, " on type ", type_name);
-            auto method_definitions_array = shared::Split(method_definitions, ',');
-            for (const shared::WSTRING& method_definition : method_definitions_array)
-            {
-                // TODO handle a * wildcard, where a * wildcard invalidates other entries for the same type
-                static const shared::WSTRING default_assembly_name = WStr("");
-                std::vector<shared::WSTRING> signatureTypes;
-                integrationDefinitions.push_back(IntegrationDefinition(
-                    MethodReference(default_assembly_name, type_name, method_definition, Version(0, 0, 0, 0), Version(USHRT_MAX, USHRT_MAX, USHRT_MAX, USHRT_MAX), signatureTypes),
-                    *trace_methods_integration_type_,
-                    false,
-                    false));
-
-                if (Logger::IsDebugEnabled())
-                {
-                    Logger::Debug("InitializeTraceMethods:  * Target: ", type_name, ".", method_definition, "(",
-                                  signatureTypes.size(), ")");
-                }
-            }
-        }
-
+        std::vector<IntegrationDefinition> integrationDefinitions = GetIntegrationsFromTraceMethodsConfiguration(dd_trace_methods, trace_methods_integration_type_);
         std::scoped_lock<std::mutex> moduleLock(module_ids_lock_);
-
-        definitions_ids_.emplace(definitionsId);
 
         Logger::Info("InitializeTraceMethods: Total number of modules to analyze: ", module_ids_.size());
         if (rejit_handler != nullptr)
