+++ tracer/src/Datadog.Trace.ClrProfiler.Native/integration.cpp
@@ -1,5 +1,6 @@
 
 #include "integration.h"
+#include "logger.h"
 
 #ifdef _WIN32
 #include <regex>
@@ -10,6 +11,7 @@
 
 #include <unordered_map>
 #include <mutex>
+#include "../../../shared/src/native-src/util.h"
 
 namespace trace
 {
@@ -37,6 +39,62 @@ AssemblyReference* AssemblyReference::GetFromCache(const shared::WSTRING& str)
     AssemblyReference* aref = new AssemblyReference(str);
     m_assemblyReferenceCache[str] = std::unique_ptr<AssemblyReference>(aref);
     return aref;
+}
+
+std::vector<IntegrationDefinition> GetIntegrationsFromTraceMethodsConfiguration(const shared::WSTRING& str,
+                                                                                TypeReference* integration_type)
+{
+    std::vector<IntegrationDefinition> integrationDefinitions;
+
+    shared::WSTRING dd_trace_methods_string = shared::WSTRING(str);
+    auto dd_trace_methods_type = shared::Split(dd_trace_methods_string, ';');
+
+    for (const shared::WSTRING& trace_method_type : dd_trace_methods_type)
+    {
+        shared::WSTRING type_name = WStr("");
+        shared::WSTRING method_definitions = WStr("");
+
+        // [] are required. Is either "*" or comma-separated values
+        // We don't know the assembly name, only the type name
+        auto firstOpenBracket = trace_method_type.find_first_of('[');
+        if (firstOpenBracket != std::string::npos)
+        {
+            auto firstCloseBracket = trace_method_type.find_first_of(']', firstOpenBracket + 1);
+            auto secondOpenBracket = trace_method_type.find_first_of('[', firstOpenBracket + 1);
+            if (firstCloseBracket != std::string::npos &&
+                (secondOpenBracket == std::string::npos || firstCloseBracket < secondOpenBracket))
+            {
+                auto length = firstCloseBracket - firstOpenBracket - 1;
+                method_definitions = trace_method_type.substr(firstOpenBracket + 1, length);
+            }
+        }
+
+        if (method_definitions.empty())
+        {
+            continue;
+        }
+
+        type_name = trace_method_type.substr(0, firstOpenBracket);
+        auto method_definitions_array = shared::Split(method_definitions, ',');
+        for (const shared::WSTRING& method_definition : method_definitions_array)
+        {
+            // TODO handle a * wildcard, where a * wildcard invalidates other entries for the same type
+            static const shared::WSTRING default_assembly_name = WStr("");
+            std::vector<shared::WSTRING> signatureTypes;
+            integrationDefinitions.push_back(IntegrationDefinition(
+                MethodReference(default_assembly_name, type_name, method_definition, Version(0, 0, 0, 0),
+                                Version(USHRT_MAX, USHRT_MAX, USHRT_MAX, USHRT_MAX), signatureTypes),
+                *integration_type, false, false));
+
+            if (Logger::IsDebugEnabled())
+            {
+                Logger::Debug("InitializeTraceMethods:  * Target: ", type_name, ".", method_definition, "(",
+                              signatureTypes.size(), ")");
+            }
+        }
+    }
+
+    return integrationDefinitions;
 }
 
 namespace
